코드 리팩토링 및 get 정리
