+++ client/src/assets/sortOpt.js
@@ -1,0 +1,15 @@
+export const planSort = [
+  { value: "dataAsc", label: "데이터양 낮은 순" },
+  { value: "dataDesc", label: "데이터양 높은 순" },
+  { value: "priceAsc", label: "요금 낮은 순" },
+  { value: "priceDesc", label: "요금 높은 순" },
+];
+
+export const planFilter = [
+  { value: "전체", label: "전체" },
+  { value: "데이터 무제한", label: "데이터 무제한" },
+  { value: "데이터 일반", label: "데이터 일반" },
+  { value: "시니어", label: "시니어" },
+  { value: "청소년", label: "청소년" },
+  { value: "다이렉트", label: "다이렉트" },
+];
