+++ client/src/components/DataOptSelect.jsx
@@ -5,8 +5,15 @@ import { useState } from "react";
 import { useEffect } from "react";
 import { getPhonesByTelecomTech } from "../api/PhoneAPI";
 import { getPlansByTelecomTech } from "../api/PlanAPI";
+import { dropdownOpt } from "../assets/dropdownOpt";
 
-export default function DataOptSelect({ label, value, handleChange, extra }) {
+export default function DataOptSelect({
+  label,
+  value,
+  handleChange,
+  extra,
+  dataList,
+}) {
   const style = {
     minWidth: 120,
     margin: "10px 0px",
@@ -32,44 +39,15 @@ export default function DataOptSelect({ label, value, handleChange, extra }) {
   const arr = () => {
     switch (label) {
       case "가입유형":
-        return [
-          { value: "번호이동", label: "번호이동" },
-          { value: "기기변경", label: "기기변경" },
-          { value: "신규가입", label: "신규가입" },
-        ];
+        return dropdownOpt().registration;
       case "할부":
-        return [
-          { value: 1, label: "완납결제" },
-          { value: 3, label: "3개월" },
-          { value: 6, label: "6개월" },
-          { value: 9, label: "9개월" },
-          { value: 10, label: "10개월" },
-          { value: 12, label: "12개월" },
-          { value: 24, label: "24개월" },
-          { value: 30, label: "30개월" },
-          { value: 36, label: "36개월" },
-          { value: 48, label: "48개월" },
-        ];
+        return dropdownOpt().installment;
       case "할인유형":
-        return extra === "다이렉트"
-          ? [{ value: 0, label: "무약정" }]
-          : [
-              { value: -1, label: "공시지원금" },
-              { value: 24, label: "선택약정24개월" },
-              { value: 12, label: "선택약정12개월" },
-            ];
+        return dropdownOpt(extra).discount;
       case "통신":
-        return [
-          { value: "5G", label: "5G" },
-          { value: "LTE", label: "LTE" },
-        ];
+        return dropdownOpt().telecomTech;
       case "제조사":
-        return [
-          { value: "전체", label: "전체" },
-          { value: "삼성", label: "삼성" },
-          { value: "애플", label: "애플" },
-          { value: "기타", label: "기타" },
-        ];
+        return dropdownOpt().manufacturingCompany;
       case "기기명":
         return list
           .filter((row) =>
