+++ client/src/components/PageContainer.jsx
@@ -3,6 +3,4 @@ import styled from "styled-components";
 export const PageContainer = styled.div`
   width: 100%;
   height: 100vh;
-  margin-left: 48px;
-  margin-right: 48px;
 `;
