+++ client/src/components/PlanModal.jsx
@@ -13,6 +13,7 @@ import sortPlanList from "../util/sortPlanList";
 import RangeSlider from "./RangeSlider";
 import { FormControl, InputLabel, MenuItem, Select } from "@mui/material";
 import PlanBox from "./PlanBox";
+import { planFilter, planSort } from "../assets/sortOpt";
 
 const style = {
   boxSizing: "border-box",
@@ -22,7 +23,7 @@ const style = {
   left: "50%",
   transform: "translate(-50%, -50%)",
   width: "1280px",
-  height: "800px",
+  height: "810px",
   bgcolor: "background.paper",
   border: "none",
   borderRadius: "10px",
@@ -53,6 +54,7 @@ export default function PlanModal({
   });
 
   const [sortBy, setSortBy] = useState("");
+  const [filterBy, setFilterBy] = useState("전체");
 
   const setRadio = (id) => setActive(id);
 
@@ -76,7 +78,57 @@ export default function PlanModal({
             Number(row.data.replace(/[^0-9.]/g, "")) >= filter.data[0] &&
             Number(row.data.replace(/[^0-9.]/g, "")) <= filter.data[1])) &&
         row.monthPrice >= filter.monthPrice[0] &&
-        row.monthPrice <= filter.monthPrice[1]
+        row.monthPrice <= filter.monthPrice[1] &&
+        ((filterBy !== "전체" && row.planType === filterBy) ||
+          filterBy === "전체")
+    );
+  };
+
+  const SelectOpt = ({ type }) => {
+    const changeValue =
+      type === "sort"
+        ? { label: "정렬", optArr: planSort, value: sortBy, handle: setSortBy }
+        : {
+            label: "유형",
+            optArr: planFilter,
+            value: filterBy,
+            handle: setFilterBy,
+          };
+    return (
+      <PlanStyle.PlanBarItem
+        style={{ width: "250px" }}
+        children={
+          <FormControl variant="standard" sx={{ minWidth: 200, height: 50 }}>
+            <InputLabel id="demo-simple-select-standard-label">
+              {changeValue.label}
+            </InputLabel>
+            <Select
+              labelId="demo-simple-select-standard-label"
+              id="demo-simple-select-standard"
+              value={changeValue.value}
+              onChange={(e) => changeValue.handle(e.target.value)}
+              label={changeValue.label}
+            >
+              {changeValue.optArr.map((row) => (
+                <MenuItem key={row.value} value={row.value}>
+                  {row.label}
+                </MenuItem>
+              ))}
+            </Select>
+          </FormControl>
+        }
+      />
+    );
+  };
+
+  const RangeOpt = ({ optKey }) => {
+    return (
+      <PlanStyle.PlanBarItem
+        style={{ width: "250px" }}
+        children={
+          <RangeSlider state={filter} setState={setFilter} optKey={optKey} />
+        }
+      />
     );
   };
 
@@ -118,64 +170,27 @@ export default function PlanModal({
             <PlanStyle.Container>
               <PlanStyle.PlanBar style={{ height: "30px", border: "none" }}>
                 <PlanStyle.PlanBarItem
-                  style={{ width: "350px" }}
+                  style={{ width: "250px" }}
                   children={<p children="요금제 정렬" />}
                 />
                 <PlanStyle.PlanBarItem
-                  style={{ width: "350px" }}
+                  style={{ width: "250px" }}
+                  children={<p children="요금제 유형" />}
+                />
+                <PlanStyle.PlanBarItem
+                  style={{ width: "250px" }}
                   children={<p children="데이터" />}
                 />
                 <PlanStyle.PlanBarItem
-                  style={{ width: "350px" }}
+                  style={{ width: "250px" }}
                   children={<p children="가격" />}
                 />
               </PlanStyle.PlanBar>
-              <PlanStyle.PlanBar style={{ height: "50px" }}>
-                <PlanStyle.PlanBarItem
-                  style={{ width: "350px" }}
-                  children={
-                    <FormControl
-                      variant="standard"
-                      sx={{ minWidth: 350, height: 50 }}
-                    >
-                      <InputLabel id="demo-simple-select-standard-label">
-                        정렬
-                      </InputLabel>
-                      <Select
-                        labelId="demo-simple-select-standard-label"
-                        id="demo-simple-select-standard"
-                        value={sortBy}
-                        onChange={(e) => setSortBy(e.target.value)}
-                        label="정렬"
-                      >
-                        <MenuItem value="dataAsc">데이터양 낮은 순</MenuItem>
-                        <MenuItem value="dataDesc">데이터양 높은 순</MenuItem>
-                        <MenuItem value="priceAsc">요금 낮은 순</MenuItem>
-                        <MenuItem value="priceDesc">요금 높은 순</MenuItem>
-                      </Select>
-                    </FormControl>
-                  }
-                />
-                <PlanStyle.PlanBarItem
-                  style={{ width: "350px" }}
-                  children={
-                    <RangeSlider
-                      state={filter}
-                      setState={setFilter}
-                      optKey="data"
-                    />
-                  }
-                />
-                <PlanStyle.PlanBarItem
-                  style={{ width: "350px" }}
-                  children={
-                    <RangeSlider
-                      state={filter}
-                      setState={setFilter}
-                      optKey="monthPrice"
-                    />
-                  }
-                />
+              <PlanStyle.PlanBar style={{ height: "60px" }}>
+                <SelectOpt type="sort" />
+                <SelectOpt type="filter" />
+                <RangeOpt optKey="data" />
+                <RangeOpt optKey="monthPrice" />
               </PlanStyle.PlanBar>
               <PlanStyle.PlanBar>
                 <PlanStyle.PlanBarItem
