+++ client/src/index.css
@@ -2,4 +2,5 @@ body {
   padding-top: 93px;
   padding-left: 48px;
   padding-right: 48px;
+  padding-bottom: 48px;
 }
