+++ client/src/styles/homeStyle.js
@@ -1,24 +1,24 @@
-import styled from 'styled-components'
+import styled from "styled-components";
 
 export const TotalLayout = styled.div`
-  padding-left : 15px;
+  padding-left: 15px;
   padding-right: 15px;
-  padding-top : 50px;
+  padding-top: 50px;
   word-wrap: break-word;
   display: flex;
   flex-direction: column;
   flex-wrap: nowrap;
-`
+`;
 export const FilterLayout = styled.div`
   width: 240px;
   margin-right: 32px;
-`
-export const FilterTitle = styled .h3`
-    display: flex;
-    justify-content: space-between;
-    height: 64px;
-    align-items: center;
-    font-weight: 400;
-    font-size: 24px;
-    padding : 10px;
-`
+`;
+export const FilterTitle = styled.h3`
+  display: flex;
+  justify-content: space-between;
+  height: 64px;
+  align-items: center;
+  font-weight: 400;
+  font-size: 24px;
+  padding: 10px;
+`;
