+++ client/src/styles/phoneListStyle.js
@@ -1,25 +1,25 @@
 import styled from "styled-components";
 
 export const InfoArea = styled.div`
-    width: 100%;
-    min-height: 48px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-`
+  width: 100%;
+  min-height: 48px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
 export const Total = styled.div`
-    height: 50px;
-    white-space: nowrap;
-`
+  height: 50px;
+  white-space: nowrap;
+`;
 export const InfoBox = styled.div`
-    display: flex;
-`
+  display: flex;
+`;
 export const Select = styled.select`
-    padding: 14px 20px 8px 0;
-    border: 1px solid transparent;
-    border-bottom-color: #bbb;
-    margin-bottom : 15px;
-`
+  padding: 14px 20px 8px 0;
+  border: 1px solid transparent;
+  border-bottom-color: #bbb;
+  margin-bottom: 15px;
+`;
 export const Option = styled.option`
-    color: #000;
-` 
+  color: #000;
+`;
