+++ client/src/styles/searchStyle.js
@@ -1,9 +1,9 @@
 import styled from "styled-components";
 
 export const TotalLayout = styled.div`
-    width: 90%;
-    overflow:hidden;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-`
+  width: 100%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
